+++ app/Rules/ValidateValue.php
@@ -17,27 +17,21 @@ class ValidateValue implements Rule
     public function passes($attribute, $value)
     {
         $db = DB::table('tm_checkarea')
-        ->select('tipe')
-        ->where('id', $this->id)->first();
-        if($db->tipe == 1)
-        {
-            if($value == 'ok' || $value == 'ng')
-            {
+            ->select('tipe')
+            ->where('id', $this->id)->first();
+
+        if ($db->tipe == 1) {
+            if ($value == 'ok' || $value == 'ng') {
                 return true;
             }
-        }
-        else if($db->tipe == 2)
-        {
+        } else if ($db->tipe == 2) {
             //if value number float big any number then true
 
-            if(is_numeric($value))
-            {
+            if (is_numeric($value)) {
                 return true;
             }
 
-        }
-        else if($db->tipe == 3)
-        {
+        } else if ($db->tipe == 3) {
             return true;
         }
