+++ resources/views/checksheet/checkdata/data.blade.php
@@ -91,105 +91,188 @@
 
 
         </div>
+        <div class="flex flex-wrap gap-2 w-full  justify-end  my-5">
+            <div class="w-full md:w-4/12    ">
+                <label for="max_tanggal">Filter</label>
+                <select id="filter" name="filter" class=" w-full border-gray-400 p-2 rounded-lg" required>
+                    <option value="" selected>Display only speciific data</option>
+                    <option value="need_check" {{ request()->get('filter') == 'need_check' ? 'selected' : '' }}>
+                        Need Check Only</option>
+                    <option value="good" {{ request()->get('filter') == 'good' ? 'selected' : '' }}>Good only</option>
+                    <option value="notgood" {{ request()->get('filter') == 'notgood' ? 'selected' : '' }}>NG Only
+                    </option>
+
+                    <option value="revised" {{ request()->get('filter') == 'revised' ? 'selected' : '' }}>Revised
+                        Only</option>
+                </select>
+            </div>
+        </div>
     </form>
     <div class="w-full overflow-x-scroll">
-        <table class="table-auto w-full table-layout-fixed border border-red-500 text-center">
-            <colgroup>
-                <col class="w-2/6">
-                <col class="w-1/12">
-                <col class="w-1/5">
-                <col class="w-2/12">
-                <col class="w-2/12">
-            </colgroup>
-            <thead class="border border-gray-400">
-                <tr>
-                    <th>Location</th>
-                    <th>Shift</th>
-                    <th>Tanggal</th>
-                    <th>Value</th>
-                    <th>Checker</th>
-                </tr>
-            </thead>
-            <tbody class="border border-gray-200 ">
+        <table class="table-auto w-full table-layout-fixed border border-red-500">
+            <tbody class="border-gray-200 border-solid">
                 @foreach ($checkdata['data'] as $data)
-                    @if ($data->status == 'good')
-                        <tr class="hover:bg-green-500 bg-green-300 border border-gray-300">
-                        @elseif($data->status == 'notgood')
-                        <tr class="hover:bg-red-500 bg-red-300 border border-gray-300">
-                    @endif
-                    <td class="py-5 border border-gray-300"><span class="text text-sm">
-                            <div>{{ $data->line }}|{{ $data->code }}</div>
-                            <div>{{ $data->nama_checksheet }}|{{ $data->nama_checkarea }}</div>
-                            <div>{{ $data->nama }}</div>
-                        </span></td>
-                    <td class="py-5 border border-gray-300">{{ $data->shift }} ({{ $data->barang }})</td>
-
-                    <td class="py-5 border border-gray-300 ">
-                        {{ $data->tanggal }}
-                    </td>
-                    <td class="py-5 border border-gray-300">
-                        <div class="flex flex-col">
-                            <div class="w-11/12 md:w-1/2 m-auto "><span
-                                    class="w-full px-5 py-1 text-gray-800 bg-gray-200 rounded-xl inline-block">{{ $data->value }}</span></div>
-                            <div class="w-11/12 md:w-1/2 m-auto  flex items-center  ">
-                                <span class=" ml-1 px-2 w-11/12 text-center py-1 text-xs font-semibold rounded-md">
-                                    {{ $data->status }}</span>
+                    <tr class="relative border border-gray-300">
+                        <td class="relative w-11/12 m-auto block my-2">
+                            <div class="flex justify-evenly border border-gray-300">
+                                <div class="w-3/12 ">
+                                    {{ $data->line }}
+                                </div>
+                                <div>
+                                    {{ $data->code }}
+                                </div>
                             </div>
-                        </div>
-
-                    </td>
-                    <td class="py-5 border border-gray-300">
-                        <div class=" flex flex-col ">
-                            <div class="w-10/12 m-auto px-2 py-1">
-
-                                <div class=" b-slate-500 text-gray-200 rounded-md bg-gray-400">
-                                    <div class="text-xs md:text-sm block mb-1">{{ $data->name }}
-                                    </div>
-                                    <div class="text-xs md:text-sm block mt-1">
-                                        ({{ $data->npk }})
+                        </td>
+                        <td class="relative w-11/12 m-auto block my-2 ">
+                            <div class="flex ">
+                                <div class="w-3/12">
+                                    Checksheet
+                                </div>
+                                <div>{{ $data->nama_checksheet }}</div>
+                            </div>
+                            <div class="flex ">
+                                <div class="w-3/12 ">
+                                    Area
+                                </div>
+                                <div>{{ $data->nama_checkarea }}</div>
+                            </div>
+                            <div class="flex ">
+                                <div class="w-3/12 ">
+                                    Cell
+                                </div>
+                                <div>{{ $data->nama }}</div>
+                            </div>
+                            <div class="flex ">
+                                <div class="w-3/12 ">
+                                    Shift
+                                </div>
+                                <div>{{ $data->shift }} ({{ $data->barang }})</div>
+                            </div>
+                            <div class="flex ">
+                                <div class="w-3/12 ">
+                                    Tanggal
+                                </div>
+                                <div>{{ $data->tanggal }}</div>
+                            </div>
+                        </td>
+                        <td class="relative w-11/12 h-auto m-auto block my-2">
+                            <div class="relative flex flex-wrap">
+                                <div class="w-3/12">
+                                    Value
+                                </div>
+                                <div class="max-h-sm flex flex-wrap gap-1">
+                                    <div class="max-h-screen">
+                                        {{ $data->value }}
+                                        <span
+                                            class="bg-gray-300 rounded-sm shadow-md shadow-black {{ $data->status == 'notgood' ? 'bg-red-300' : 'bg-green-300' }}  py-0.5 px-5">{{ $data->status == 'notgood' ? 'NG' : 'OK' }}</span>
                                     </div>
                                 </div>
-
-
-
                             </div>
-                            @if (auth()->user()->role == 'admin')
-                                <div class="w-3/4 md:w-3/4 m-auto flex items-center  ">
-                                    @include('checksheet.checkdata.changeStatus')
+                        </td>
+                        @if ($data->notes != null || $data->notes != '')
+                            <?php
+                            $data->notes = explode('<br>', $data->notes)[0];
+                            $data->notes = htmlspecialchars($data->notes);
+
+                            ?>
+                            <td class="relative w-11/12 h-auto m-auto block my-2">
+                                <div class="relative flex flex-wrap">
+                                    <div class="w-3/12">
+                                        Notes
+                                    </div>
+                                    <div class="max-h-sm flex flex-wrap gap-1">
+                                        <div class="max-h-screen max-w-full py-0.5 pr-2 bg-gray-100 rounded-sm">
+                                            {{ $data->notes }}
+                                        </div>
+                                        @if ($data->mark != null || $data->mark != '')
+                                            <div>
+                                                @if ($data->mark == 1)
+                                                    <span
+                                                        class=" rounded-sm shadow-md shadow-black bg-green-300  py-0.5 px-5">Problem
+                                                        Solved</span>
+                                                @endif
+                                            </div>
+                                        @endif
+                                    </div>
                                 </div>
-                            @else
-                                @if ($data->approval == '4')
-                                    <div class="w-3/4 md:w-1/2m-auto flex items-center">
-                                        <span
-                                            class="inline-block ml-1 px-2 py-1 w-11/12 text-center text-sm font-semibold text-gray-800 bg-green-200 rounded-md">
-                                            Approved
-                                        </span>
+                            </td>
+                            @if ($data->mark != null || $data->mark == 1)
+                                <td class="relative w-11/12 h-auto m-auto block my-2">
+                                    <div class="relative flex flex-wrap">
+                                        <div class="w-3/12">
+                                            Revised Value
+                                        </div>
+                                        <div class="max-h-sm flex flex-wrap gap-1">
+                                            <div class="max-h-screen">
+                                                {{ $data->revised_value }}
+                                                <span
+                                                    class="bg-gray-300 rounded-sm shadow-md shadow-black {{ $data->revised_status == 'notgood' ? 'bg-red-300' : 'bg-green-300' }}  py-0.5 px-5">{{ $data->revised_status == 'notgood' ? 'NG' : 'OK' }}</span>
+                                            </div>
+                                        </div>
                                     </div>
-                                @elseif($data->approval < 1 && $data->approval > 4)
-                                    <div class="w-3/4 md:w-1/2 m-auto flex items-center">
-                                        <span
-                                            class="inline-block ml-1 px-2 py-1 w-11/12 text-center text-sm font-semibold text-gray-800 bg-red-200 rounded-md">
-                                            Rejected
-                                        </span>
-                                    </div>
-                                @else
-                                    <div class="w-3/4 md:w-1/2 m-auto flex items-center ">
-                                        <span
-                                            class="inline-block ml-1 px-2 w-11/12 text-center py-1 text-sm font-semibold text-gray-800 bg-yellow-200 rounded-md">
-                                            Waiting
-                                        </span>
-                                    </div>
-                                @endif
+                                </td>
                             @endif
-
-                        </div>
-                    </td>
+                        @endif
+                        <td class="relative w-11/12 h-auto m-auto block my-2">
+                            <div class="relative flex flex-wrap">
+                                <div class="w-3/12">
+                                    Checker
+                                </div>
+                                <div class="max-h-sm flex flex-wrap gap-1">
+                                    <div class="max-h-screen">
+                                        {{ $data->name }}
+                                        <span
+                                            class="bg-gray-300 rounded-sm shadow-md shadow-black  py-0.5 px-5">{{ $data->npk }}</span>
+                                    </div>
+                                </div>
+                            </div>
+                        </td>
+                        <td class="relative w-11/12 h-auto m-auto block my-2">
+                            <div class="relative flex flex-wrap">
+                                <div class="w-3/12">
+                                    Approval
+                                </div>
+                                <div class=" w-full flex flex-wrap justify-between border border-gray-300 py-2 px-5">
+                                    <div class="max-h-screen flex flex-col gap-5  items-center text-center">
+                                        <div class="font-bold text-lg">GP</div>
+                                        @if ($data->approval >= 1)
+                                            <div class="rounded-full w-5 h-5 bg-green-300"></div>
+                                        @else
+                                            <div class="rounded-full w-5 h-5 bg-gray-300"></div>
+                                        @endif
+                                    </div>
+                                    <div class="max-h-screen flex flex-col gap-5 items-center text-center">
+                                        <div class="font-bold text-lg">Leader</div>
+                                        @if ($data->approval >= 2)
+                                            <div class="rounded-full w-5 h-5 bg-green-300"></div>
+                                        @else
+                                            <div class="rounded-full w-5 h-5 bg-gray-300"></div>
+                                        @endif
+                                    </div>
+                                    <div class="max-h-screen flex flex-col gap-5  items-center text-center">
+                                        <div class="font-bold text-lg">Supervisor</div>
+                                        @if ($data->approval >= 3)
+                                            <div class="rounded-full w-5 h-5 bg-green-300"></div>
+                                        @else
+                                            <div class="rounded-full w-5 h-5 bg-gray-300"></div>
+                                        @endif
+                                    </div>
+                                    <div class="max-h-screen flex flex-col gap-5  items-center text-center">
+                                        <div class="font-bold text-lg">Manager</div>
+                                        @if ($data->approval >= 4)
+                                            <div class="rounded-full w-5 h-5 bg-green-300"></div>
+                                        @else
+                                            <div class="rounded-full w-5 h-5 bg-gray-300"></div>
+                                        @endif
+                                    </div>
+                                </div>
+                            </div>
+                        </td>
                     </tr>
                 @endforeach
             </tbody>
         </table>
     </div>
-
 @endsection
 @section('script')
     <script>
@@ -209,7 +292,7 @@
         @if (auth()->user()->role == 'admin')
 
 
-            function changeStatus(element){
+            function changeStatus(element) {
                 var id = element.getAttribute('data-id');
                 var status = element.value;
                 var url = "{{ route('checksheet.data.changeStatus', ':id') }}";
@@ -241,8 +324,7 @@
                     .catch((error) => {
                         console.error('Error:', error);
                     });
-                }
-
+            }
         @endif
     </script>
 @endsection
