+++ resources/views/layouts/navigation.blade.php
@@ -20,12 +20,10 @@
                         {{ __('Checksheet List') }}
 
                     </x-nav-link>
-                    @if(auth()->user()->role == "admin")
                     <x-nav-link :href="route('checksheet.data')" :active="request()->routeIs('checksheet.data')">
                         {{ __('Checksheet Data') }}
 
                     </x-nav-link>
-                    @endif
 
                 </div>
             </div>
@@ -90,12 +88,10 @@
                 {{ __('Checksheet List') }}
 
             </x-responsive-nav-link>
-            @if(auth()->user()->role == "admin")
             <x-responsive-nav-link :href="route('checksheet.data')" :active="request()->routeIs('checksheet.data')">
                 {{ __('Checksheet Data') }}
 
             </x-responsive-nav-link>
-            @endif
         </div>
 
         <!-- Responsive Settings Options -->
